se mofico entidad de acceso a tabla usuariolog
