+++ sigpro/src/pojo/Usuariolog.java
@@ -1,12 +1,13 @@
 package pojo;
-// Generated 16/12/2016 09:52:23 AM by Hibernate Tools 5.2.0.Beta1
+// Generated 16/12/2016 03:14:50 PM by Hibernate Tools 5.2.0.Beta1
 
-import javax.persistence.AttributeOverride;
-import javax.persistence.AttributeOverrides;
+import java.util.Date;
 import javax.persistence.Column;
-import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 
 /**
  * Usuariolog generated by hbm2java
@@ -18,26 +19,37 @@ public class Usuariolog implements java.io.Serializable {
 	/**
 	 * 
 	 */
-	private static final long serialVersionUID = -5762512060655483094L;
-	private UsuariologId id;
+	private static final long serialVersionUID = -1163268594466555337L;
+	private String usuario;
+	private Date fecha;
 
 	public Usuariolog() {
 	}
 
-	public Usuariolog(UsuariologId id) {
-		this.id = id;
+	public Usuariolog(String usuario, Date fecha) {
+		this.usuario = usuario;
+		this.fecha = fecha;
 	}
 
-	@EmbeddedId
+	@Id
 
-	@AttributeOverrides({ @AttributeOverride(name = "usuario", column = @Column(name = "usuario", length = 50)),
-			@AttributeOverride(name = "fecha", column = @Column(name = "fecha", nullable = false, length = 19)) })
-	public UsuariologId getId() {
-		return this.id;
+	@Column(name = "usuario", unique = true, nullable = false, length = 32)
+	public String getUsuario() {
+		return this.usuario;
 	}
 
-	public void setId(UsuariologId id) {
-		this.id = id;
+	public void setUsuario(String usuario) {
+		this.usuario = usuario;
+	}
+
+	@Temporal(TemporalType.TIMESTAMP)
+	@Column(name = "fecha", nullable = false, length = 19)
+	public Date getFecha() {
+		return this.fecha;
+	}
+
+	public void setFecha(Date fecha) {
+		this.fecha = fecha;
 	}
 
 }
