--- sigpro/src/pojo/UsuariologId.java
@@ -1,74 +1,0 @@
-package pojo;
-// Generated 16/12/2016 09:52:23 AM by Hibernate Tools 5.2.0.Beta1
-
-import java.util.Date;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-
-/**
- * UsuariologId generated by hbm2java
- */
-@Embeddable
-public class UsuariologId implements java.io.Serializable {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -4437826725793738708L;
-	private String usuario;
-	private Date fecha;
-
-	public UsuariologId() {
-	}
-
-	public UsuariologId(Date fecha) {
-		this.fecha = fecha;
-	}
-
-	public UsuariologId(String usuario, Date fecha) {
-		this.usuario = usuario;
-		this.fecha = fecha;
-	}
-
-	@Column(name = "usuario", length = 50)
-	public String getUsuario() {
-		return this.usuario;
-	}
-
-	public void setUsuario(String usuario) {
-		this.usuario = usuario;
-	}
-
-	@Column(name = "fecha", nullable = false, length = 19)
-	public Date getFecha() {
-		return this.fecha;
-	}
-
-	public void setFecha(Date fecha) {
-		this.fecha = fecha;
-	}
-
-	public boolean equals(Object other) {
-		if ((this == other))
-			return true;
-		if ((other == null))
-			return false;
-		if (!(other instanceof UsuariologId))
-			return false;
-		UsuariologId castOther = (UsuariologId) other;
-
-		return ((this.getUsuario() == castOther.getUsuario()) || (this.getUsuario() != null
-				&& castOther.getUsuario() != null && this.getUsuario().equals(castOther.getUsuario())))
-				&& ((this.getFecha() == castOther.getFecha()) || (this.getFecha() != null
-						&& castOther.getFecha() != null && this.getFecha().equals(castOther.getFecha())));
-	}
-
-	public int hashCode() {
-		int result = 17;
-
-		result = 37 * result + (getUsuario() == null ? 0 : this.getUsuario().hashCode());
-		result = 37 * result + (getFecha() == null ? 0 : this.getFecha().hashCode());
-		return result;
-	}
-
-}
